Proyectos/p2/requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
kiwisolver==1.4.9
markdown-it-py==4.0.0
matplotlib==3.10.7
mdurl==0.1.2
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
rich==14.2.0
seaborn==0.13.2
six==1.17.0
tabulate==0.9.0
tqdm==4.67.1
tzdata==2025.2
